requirements.txt
fastmcp==2.9.2
psycopg2-binary==2.9.10
uvicorn
